Import CSV file to MySQL Database
There are mixed data types in 'Perpetrator Age' column. Also, one of the rows has space character in the form of string data type. Due to this issue, the code will throw an error during the import. The code I have uploaded will not throw an error and uploads all the 638454 rows in the database because I have cleaned the data. The cleaned data is available in the other file, "HomicideReports_Data_Prep&Viz.py".